fnbill_mock_backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
reportlab
python-dotenv